requirements.txt
# Real requirements
python-telegram-bot==11.1.0
beautifulsoup4==4.7.1
Flask==1.0.2

# Dependencies
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
Click==7.0
cryptography==2.6.1
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pycparser==2.19
six==1.12.0
soupsieve==1.8
Werkzeug==0.14.1
